Move field name requirements to recommendations.
Field names can be opaque to the spec, much like naming rules for 
resource types and URLs. By keeping all naming rules as recommendations,
the spec is leaner and more flexible.

The naming rules did mitigate a potential conflict with default sort
order, in which names couldn’t begin with a `-`. As an alternative,
the concept of a default order is removed; sort order must now be
explicitly specified with a `+` or `-` prefix.